requirements.txt
flask>=3.0,<4
groq>=0.11,<1
gTTS>=2.5,<3
gunicorn[gevent]>=22,<24
python-dotenv>=1.0,<2
trafilatura>=1.12,<2
